refactor: move package orderers to plugin system (rebased from #1787)
Move the five built-in package orderers (NullPackageOrder, SortedOrder,
PerFamilyOrder, VersionSplitPackageOrder, TimestampPackageOrder) from
src/rez/package_order.py into src/rezplugins/package_order/ as proper
rez plugins.

Register PackageOrderPluginType in plugin_managers.py.

Replace PR #1787's function shims with lazy module-level __getattr__
aliases (PEP 562) so isinstance, from_pod, and type(x) is checks all
work, while avoiding circular imports at module load time.

Keep _orderers dict and register_orderer() as legacy fallback, marked
with deprecation comments.

Original work by Robert Minsk (PR #1787). Rebased, conflicts resolved,
and fixed by Stephen Mackenzie.
