+++ src/rez/package_order.py
@@ -9,10 +9,11 @@ from hashlib import sha1
 from typing import Any, Callable, Iterable, List, TYPE_CHECKING
 
 from rez.config import config
+from rez.exceptions import RezPluginError
 from rez.utils.data_utils import cached_class_property
 from rez.version import Version, VersionRange
-from rez.version._version import _Comparable, _ReversedComparable, _LowerBound, _UpperBound, _Bound
-from rez.packages import iter_packages, Package
+from rez.version._version import _Comparable, _LowerBound, _UpperBound, _Bound
+from rez.packages import Package
 from rez.utils.typing import SupportsLessThan
 
 if TYPE_CHECKING:
@@ -193,7 +194,7 @@ class PackageOrder(object):
         raise NotImplementedError
 
     def __eq__(self, other):
-        return type(self) == type(other) and str(self) == str(other)  # noqa: E721
+        return type(self) is type(other) and str(self) == str(other)
 
     def __ne__(self, other) -> bool:
         return not self == other
@@ -202,434 +203,51 @@ class PackageOrder(object):
         return "%s(%s)" % (self.__class__.__name__, str(self))
 
 
-class NullPackageOrder(PackageOrder):
-    """An orderer that does not change the order - a no op.
+# Legacy orderer registry. Used as fallback by _find_orderer when an orderer
+# is not found in the plugin system.
+_orderers = {}
 
-    This orderer is useful in cases where you want to apply some default orderer
-    to a set of packages, but may want to explicitly NOT reorder a particular
-    package. You would use a :class:`NullPackageOrder` in a :class:`PerFamilyOrder` to do this.
+
+def _find_orderer(name):
+    """Find an orderer class by name.
+
+    Checks the plugin system first, then falls back to the legacy _orderers
+    registry for backward compatibility with register_orderer().
     """
-    name = "no_order"
+    from rez.plugin_managers import plugin_manager
 
-    def sort_key_implementation(self, package_name: str, version: Version) -> SupportsLessThan:
-        # python's sort will preserve the order of items that compare equal, so
-        # to not change anything, we just return the same object for all...
-        return 0
-
-    def __str__(self) -> str:
-        return "{}"
-
-    def __eq__(self, other):
-        return type(self) == type(other)  # noqa: E721
-
-    def to_pod(self) -> dict[str, Any]:
-        """
-        Example (in yaml):
-
-        .. code-block:: yaml
-
-           type: no_order
-           packages: ["foo"]
-        """
-        return {
-            "packages": self.packages,
-        }
-
-    @classmethod
-    def from_pod(cls, data: dict[str, Any]) -> Self:
-        return cls(packages=data.get("packages"))
+    try:
+        return plugin_manager.get_plugin_class('package_order', name)
+    except RezPluginError:
+        # Fallback to legacy "register_orderer" method
+        if name not in _orderers:
+            raise
+        return _orderers[name]
 
 
-class SortedOrder(PackageOrder):
-    """An orderer that sorts based on :attr:`Package.version <rez.packages.Package.version>`.
+# Backward-compat aliases. These are resolved lazily via __getattr__ below to
+# avoid a circular import: the plugin modules in rezplugins/package_order/
+# import from this module, so we cannot call _find_orderer (which loads the
+# plugin system) at module load time.
+#
+# New code should import from the plugin system directly.
+_LEGACY_ORDERER_NAMES = {
+    "NullPackageOrder": "no_order",
+    "PerFamilyOrder": "per_family",
+    "SortedOrder": "sorted",
+    "TimestampPackageOrder": "soft_timestamp",
+    "VersionSplitPackageOrder": "version_split",
+}
+
+
+def __getattr__(name):
+    """Provide backward-compat class aliases for orderers now in the plugin system.
+
+    Resolved lazily to avoid circular imports at module load time.
     """
-    name = "sorted"
-
-    def __init__(self, descending: bool, packages: list[str] | None = None) -> None:
-        super().__init__(packages)
-        self.descending = descending
-
-    def sort_key_implementation(self, package_name: str, version: Version) -> SupportsLessThan:
-        # Note that the name "descending" can be slightly confusing - it
-        # indicates that the final ordering this Order gives should be
-        # version descending (ie, the default) - however, the sort_key itself
-        # returns its results in "normal" ascending order (because it needs to
-        # be used "alongside" normally-sorted objects like versions).
-        # when the key is passed to sort(), though, it is always invoked with
-        # reverse=True...
-        if self.descending:
-            return version
-        else:
-            return _ReversedComparable(version)
-
-    def __str__(self) -> str:
-        return str(self.descending)
-
-    def __eq__(self, other):
-        return (  # noqa: E721
-            type(self) == type(other)
-            and self.descending == other.descending
-        )
-
-    def to_pod(self) -> dict[str, Any]:
-        """
-        Example (in yaml):
-
-        .. code-block:: yaml
-
-           type: sorted
-           descending: true
-           packages: ["foo"]
-        """
-        return {
-            "descending": self.descending,
-            "packages": self.packages,
-        }
-
-    @classmethod
-    def from_pod(cls, data: dict[str, Any]) -> Self:
-        return cls(
-            data["descending"],
-            packages=data.get("packages"),
-        )
-
-
-class PerFamilyOrder(PackageOrder):
-    """An orderer that applies different orderers to different package families.
-    """
-    name = "per_family"
-
-    def __init__(self, order_dict: dict[str, PackageOrder],
-                 default_order: PackageOrder | None = None) -> None:
-        """Create a reorderer.
-
-        Args:
-            order_dict (dict[str, PackageOrder]): Orderers to apply to
-                each package family.
-            default_order (PackageOrder): Orderer to apply to any packages
-                not specified in ``order_dict``.
-        """
-        super().__init__(list(order_dict))
-        self.order_dict = order_dict.copy()
-        self.default_order = default_order
-
-    def reorder(self, iterable: Iterable[Package],
-                key: Callable[[Any], Package] | None = None) -> list[Package] | None:
-        package_name = self._get_package_name_from_iterable(iterable, key)
-        if package_name is None:
-            return None
-
-        orderer = self.order_dict.get(package_name)
-        if orderer is None:
-            orderer = self.default_order
-        if orderer is None:
-            return None
-
-        return orderer.reorder(iterable, key)
-
-    def sort_key_implementation(self, package_name: str, version: Version) -> SupportsLessThan:
-        orderer = self.order_dict.get(package_name)
-        if orderer is None:
-            if self.default_order is None:
-                # shouldn't get here, because applies_to should protect us...
-                raise RuntimeError(
-                    "package family orderer %r does not apply to package family %r",
-                    (self, package_name))
-
-            orderer = self.default_order
-
-        return orderer.sort_key_implementation(package_name, version)
-
-    def __str__(self) -> str:
-        items = sorted((x[0], str(x[1])) for x in self.order_dict.items())
-        return str((items, str(self.default_order)))
-
-    def __eq__(self, other):
-        return (  # noqa: E721
-            type(other) == type(self)
-            and self.order_dict == other.order_dict
-            and self.default_order == other.default_order
-        )
-
-    def to_pod(self) -> dict[str, Any]:
-        """
-        Example (in yaml):
-
-        .. code-block:: yaml
-
-           type: per_family
-           orderers:
-           - packages: ['foo', 'bah']
-             type: version_split
-             first_version: '4.0.5'
-           - packages: ['python']
-             type: sorted
-             descending: false
-           default_order:
-             type: sorted
-             descending: true
-        """
-        orderers = {}
-        packages = {}
-
-        # group package fams by orderer they use
-        for fam, orderer in self.order_dict.items():
-            k = id(orderer)
-            orderers[k] = orderer
-            packages.setdefault(k, set()).add(fam)
-
-        orderlist = []
-        for k, fams in packages.items():
-            orderer = orderers[k]
-            data = to_pod(orderer)
-            data["packages"] = sorted(fams)
-            orderlist.append(data)
-
-        result: dict[str, Any] = {"orderers": orderlist}
-
-        if self.default_order is not None:
-            result["default_order"] = to_pod(self.default_order)
-
-        return result
-
-    @classmethod
-    def from_pod(cls, data: dict[str, Any]) -> Self:
-        order_dict = {}
-        default_order = None
-
-        for d in data["orderers"]:
-            d = d.copy()
-            fams = d.pop("packages")
-            orderer = from_pod(d)
-
-            for fam in fams:
-                order_dict[fam] = orderer
-
-        d = data.get("default_order")
-        if d:
-            default_order = from_pod(d)
-
-        return cls(order_dict, default_order)
-
-
-class VersionSplitPackageOrder(PackageOrder):
-    """Orders package versions <= a given version first.
-
-    For example, given the versions [5, 4, 3, 2, 1], an orderer initialized
-    with ``version=3`` would give the order [3, 2, 1, 5, 4].
-    """
-    name = "version_split"
-
-    def __init__(self, first_version: Version, packages: list[str] | None = None) -> None:
-        """Create a reorderer.
-
-        Args:
-            first_version (Version): Start with versions <= this value.
-        """
-        super().__init__(packages)
-        self.first_version = first_version
-
-    def sort_key_implementation(self, package_name: str, version: Version) -> SupportsLessThan:
-        priority_key = 1 if version <= self.first_version else 0
-        return priority_key, version
-
-    def __str__(self) -> str:
-        return str(self.first_version)
-
-    def __eq__(self, other):
-        return (  # noqa: E721
-            type(other) == type(self)
-            and self.first_version == other.first_version
-        )
-
-    def to_pod(self) -> dict[str, Any]:
-        """
-        Example (in yaml):
-
-        .. code-block:: yaml
-
-           type: version_split
-           first_version: "3.0.0"
-           packages: ["foo"]
-        """
-        return dict(
-            first_version=str(self.first_version),
-            packages=self.packages,
-        )
-
-    @classmethod
-    def from_pod(cls, data: dict[str, Any]) -> Self:
-        return cls(
-            Version(data["first_version"]),
-            packages=data.get("packages"),
-        )
-
-
-class TimestampPackageOrder(PackageOrder):
-    """A timestamp order function.
-
-    Given a time ``T``, this orderer returns packages released before ``T``, in descending
-    order, followed by those released after. If ``rank`` is non-zero, version
-    changes at that rank and above are allowed over the timestamp.
-
-    For example, consider the common case where we want to prioritize packages
-    released before ``T``, except for newer patches. Consider the following package
-    versions, and time ``T``:
-
-    .. code-block:: text
-
-       2.2.1
-       2.2.0
-       2.1.1
-       2.1.0
-       2.0.6
-       2.0.5
-             <-- T
-       2.0.0
-       1.9.0
-
-    A timestamp orderer set to ``rank=3`` (patch versions) will attempt to consume
-    the packages in the following order:
-
-    .. code-block:: text
-
-       2.0.6
-       2.0.5
-       2.0.0
-       1.9.0
-       2.1.1
-       2.1.0
-       2.2.1
-       2.2.0
-
-    Notice that packages before ``T`` are preferred, followed by newer versions.
-    Newer versions are consumed in ascending order, except within rank (this is
-    why 2.1.1 is consumed before 2.1.0).
-    """
-    name = "soft_timestamp"
-
-    def __init__(self, timestamp: int, rank: int = 0, packages: list[str] | None = None) -> None:
-        """Create a reorderer.
-
-        Args:
-            timestamp (int): Epoch time of timestamp. Packages before this time
-                are preferred.
-            rank (int): If non-zero, allow version changes at this rank or above
-                past the timestamp.
-        """
-        super().__init__(packages)
-        self.timestamp = timestamp
-        self.rank = rank
-
-        # dictionary mapping from package family to the first-version-after
-        # the given timestamp
-        self._cached_first_after = {}
-        self._cached_sort_key = {}
-
-    def _get_first_after(self, package_family: str) -> Version | None:
-        """Get the first package version that is after the timestamp"""
-        try:
-            first_after = self._cached_first_after[package_family]
-        except KeyError:
-            first_after = self._calc_first_after(package_family)
-            self._cached_first_after[package_family] = first_after
-        return first_after
-
-    def _calc_first_after(self, package_family: str) -> Version | None:
-        descending = sorted(iter_packages(package_family),
-                            key=lambda p: p.version,
-                            reverse=True)
-        first_after = None
-        for i, package in enumerate(descending):
-            if not package.timestamp:
-                continue
-            if package.timestamp > self.timestamp:
-                first_after = package.version
-            else:
-                break
-
-        if not self.rank:
-            return first_after
-
-        # if we have rank, then we need to then go back UP the
-        # versions, until we find one whose trimmed version doesn't
-        # match.
-        # Note that we COULD do this by simply iterating through
-        # an ascending sequence, in which case we wouldn't have to
-        # "switch direction" after finding the first result after
-        # by timestamp... but we're making the assumption that the
-        # timestamp break will be closer to the higher end of the
-        # version, and that we'll therefore have to check fewer
-        # timestamps this way...
-        trimmed_version = package.version.trim(self.rank - 1)
-        first_after = None
-        for after_package in reversed(descending[:i]):
-            if after_package.version.trim(self.rank - 1) != trimmed_version:
-                return after_package.version
-
-        return first_after
-
-    def _calc_sort_key(self, package_name: str, version: Version) -> SupportsLessThan:
-        first_after = self._get_first_after(package_name)
-        if first_after is None:
-            # all packages are before T
-            is_before: bool | int = True
-        else:
-            is_before = int(version < first_after)
-
-        if is_before:
-            return is_before, version
-
-        if self.rank:
-            return (is_before,
-                    _ReversedComparable(version.trim(self.rank - 1)),
-                    version.tokens[self.rank - 1:])
-
-        return is_before, _ReversedComparable(version)
-
-    def sort_key_implementation(self, package_name: str, version: Version) -> SupportsLessThan:
-        cache_key = (package_name, str(version))
-        result = self._cached_sort_key.get(cache_key)
-        if result is None:
-            result = self._calc_sort_key(package_name, version)
-            self._cached_sort_key[cache_key] = result
-
-        return result
-
-    def __str__(self) -> str:
-        return str((self.timestamp, self.rank))
-
-    def __eq__(self, other):
-        return (  # noqa: E721
-            type(other) == type(self)
-            and self.timestamp == other.timestamp
-            and self.rank == other.rank
-        )
-
-    def to_pod(self) -> dict[str, Any]:
-        """
-        Example (in yaml):
-
-        .. code-block:: yaml
-
-           type: soft_timestamp
-           timestamp: 1234567
-           rank: 3
-           packages: ["foo"]
-        """
-        return dict(
-            timestamp=self.timestamp,
-            rank=self.rank,
-            packages=self.packages,
-        )
-
-    @classmethod
-    def from_pod(cls, data: dict[str, Any]) -> Self:
-        return cls(
-            data["timestamp"],
-            rank=data.get("rank", 0),
-            packages=data.get("packages"),
-        )
+    if name in _LEGACY_ORDERER_NAMES:
+        return _find_orderer(_LEGACY_ORDERER_NAMES[name])
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
 class PackageOrderList(List[PackageOrder]):
@@ -656,6 +274,19 @@ class PackageOrderList(List[PackageOrder]):
     def singleton(cls) -> PackageOrderList:
         """Filter list as configured by rezconfig.package_filter."""
         return cls.from_pod(config.package_orderers)
+
+    @classmethod
+    def clear_singleton_cache(cls) -> None:
+        """Clear the cached singleton so the next access re-reads config.
+
+        Use this when runtime configuration (e.g. REZ_PACKAGE_ORDERERS_JSON)
+        has changed and you want the new config to take effect. Note that
+        the config-level cache must also be cleared via
+        ``config._uncache("package_orderers")``.
+        """
+        name = "_class_property_singleton"
+        if hasattr(cls, name):
+            delattr(cls, name)
 
     @staticmethod
     def _to_orderer(orderer: dict | PackageOrder) -> PackageOrder:
@@ -727,12 +358,12 @@ def from_pod(data: dict[str, Any]) -> PackageOrder:
         data = data.copy()
         data.pop("type")
 
-        cls = _orderers[cls_name]
+        cls = _find_orderer(cls_name)
         return cls.from_pod(data)
     else:
         # old-style, kept for backwards compatibility
         cls_name, data_ = data
-        cls = _orderers[cls_name]
+        cls = _find_orderer(cls_name)
         return cls.from_pod(data_)
 
 
@@ -744,15 +375,26 @@ def get_orderer(package_name: str, orderers: PackageOrderList | dict[str, Packag
         orderer = orderers.get(ALL_PACKAGES)
     if orderer is None:
         # default ordering is version descending
-        orderer = SortedOrder(descending=True)
+        sorted_order = _find_orderer("sorted")
+        orderer = sorted_order(descending=True)
     return orderer
 
 
+# Legacy orderer registry. Used as fallback by _find_orderer when an orderer
+# is not found in the plugin system.
+_orderers = {}
+
+
+# Legacy orderer registration. New orderers should be implemented as plugins
+# in rezplugins/package_order/. This function is kept for backward
+# compatibility with custom orderers registered via rezconfig.py.
 def register_orderer(cls: type[PackageOrder]) -> bool:
-    """Register an orderer
+    """Register an orderer.
+
+    Kept for backwards compatibility. New orderers should be a plugin.
 
     Args:
-        cls (type[PackageOrder]): Package orderer class to register.
+        cls (type[PackageOrder]): Package order class to register.
 
     returns:
         bool: True if successfully registered, else False.
@@ -763,9 +405,3 @@ def register_orderer(cls: type[PackageOrder]) -> bool:
         return True
     else:
         return False
-
-
-# registration of builtin orderers
-_orderers = {}
-for o in list(globals().values()):
-    register_orderer(o)
